requirement.txt
altgraph==0.17.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
asyncio==3.4.3
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blobfile==2.0.1
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.5
cron-lite==1.0
croniter==1.3.8
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.0.0rc9
executing==1.2.0
ExifRead==3.0.0
fastjsonschema==2.16.2
filelock==3.9.0
Flask==2.2.2
fonttools==4.37.3
fqdn==1.5.1
future==0.18.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
Nuitka==1.4.8
numpy==1.23.3
OpenAIAuth==0.3.2
outcome==1.2.0
packaging==21.3
pandas==1.5.0
pandocfilters==1.5.0
parso==0.8.3
pefile==2022.5.30
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.17
Pygments==2.13.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
PyMySQL==1.0.2
pyparsing==3.0.9
pyperclip==1.8.2
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.19.2
PySide2==5.15.2.1
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.2.1
pywin32==305
pywin32-ctypes==0.2.0
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.2
revChatGPT==3.0.9
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
selenium==4.5.0
Send2Trash==1.8.0
shiboken2==5.15.2.1
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tencentcloud-sdk-python==3.0.755
tencentcloud-sdk-python-dnspod==3.0.756
terminado==0.17.1
tiktoken==0.3.0
tinycss2==1.2.1
tornado==6.2
traitlets==5.7.0
trio==0.22.0
trio-websocket==0.9.2
uri-template==1.2.0
urllib3==1.26.14
wcwidth==0.2.6
webcolors==1.12
webdav4==0.9.8
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
wsproto==1.2.0
yarg==0.1.9
zipp==3.10.0
[0m